user-accounts: Drop "Restart" notification
We have code to show the "Restart" notification when the user
would change its account-type between STANDARD/ADMIN, except that
we never allow that to happen. So this code was never used.

We have the would_demote_only_admin function that is used to set
the visibility of the Account Type row. So we won't ever display
the Account Type row if the user would demote the only admin.